apps/server/pyproject.toml
[tool.poetry]
name = "jouster-server"
version = "0.1.0"
description = ""
authors = ["Brian B"]
readme = "README.md"
# ------------------------------------------------------------------
# Add this line to explicitly tell Poetry to not treat this as a package.
package-mode = false
# ------------------------------------------------------------------

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.116.1"
uvicorn = { extras = ["standard"], version = "^0.35.0" }
prisma = "^0.15.0"
psycopg2-binary = "^2.9.10"
httpx = "^0.28.1"
openai = "^1.107.1"
spacy = "^3.8.7"
spacy-lookups-data = "^1.0.5"
loguru = "^0.7.3"
pydantic-settings = "^2.10.1"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
